corrected uniform setting code
using projection model view matrices
fixed warning about float casting

+++ src/main.cpp
@@ -108,7 +108,7 @@ int main(int argc, char *argv[]) {
 
         mat4 model = rotate(mat4(1.0f), radians(-90.0f), vec3(1.0f, 0.0f, 0.0f));
         const mat4 view = translate(mat4(1.0f), vec3(0.0f, 0.0f, -3.0f));
-        const mat4 projection = perspective(radians(45.0f), static_cast<float>(window_w / window_h), 0.1f, 100.0f);
+        const mat4 projection = perspective(radians(45.0f), (float)window_w / (float)window_h, 0.1f, 100.0f);
 
         program.use();
         program.set_offset_x(0.0);
